Added Controller, Services, Entity and Repository

+++ src/main/java/com/example/LibraryManagementSystem/repository/AuthorRepo.java
@@ -3,5 +3,5 @@ package com.example.LibraryManagementSystem.repository;
 import com.example.LibraryManagementSystem.entity.Author;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuthorRepo extends JpaRepository<Author, Integer> {
+public interface AuthorRepo extends JpaRepository<Author, Long> {
 }
