+++ src/main/java/com/example/LibraryManagementSystem/repository/BookRepo.java
@@ -3,5 +3,5 @@ package com.example.LibraryManagementSystem.repository;
 import com.example.LibraryManagementSystem.entity.Books;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepo extends JpaRepository<Books, Integer> {
+public interface BookRepo extends JpaRepository<Books, Long> {
 }
