+++ src/main/java/com/example/LibraryManagementSystem/repository/CategoryRepo.java
@@ -3,5 +3,5 @@ package com.example.LibraryManagementSystem.repository;
 import com.example.LibraryManagementSystem.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepo extends JpaRepository<Category, Integer> {
+public interface CategoryRepo extends JpaRepository<Category, Long> {
 }
